+++ src/Commands/Languages/Cypher/CommandProcessor.php
@@ -56,7 +56,7 @@ class CommandProcessor implements ProcessorInterface
      * script for whichever driver is specified
      *
      * @param Bag $bag
-     * @return CommandInterface
+     * @return \Spider\Commands\Command
      */
     public function process(Bag $bag)
     {