+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All Notable changes to `Spider` will be documented in this file
 
+## v0.4 - NEXT
+- Query api uses verb-based dispatch methods (getAll(), getOne(), etc)
+- Refactor: OrientDB uses SqlBatch for commands consistently
+- Simplified Drivers by replacing read/writeCommand() with single executeCommand() and runCommand()
+- Adds a Vagrant box for easier development environment.
+
 ## v0.3.2 - 10-8-2015
 - Fixes dependency versions in composer.json
 - Fixes bug #91: Gremlin-PHP updated version was breaking change
@@ -8,6 +14,7 @@ All Notable changes to `Spider` will be documented in this file
 ## v0.3.1 - 9-1-2015
 - Fixes bug #63: Set ordering in tests for Neo4j
 - Cleans up scrutinizer
+- Bug: Added ordering to select tests so Neo would consistently pass #63
 
 ## v0.3.0 - 8-22-2015
 - Basic Query Builder (without traversals)